station-sites-management/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
Flask==3.0.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pymongo==4.6.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
